[FIX] account: fixes reconciliation mode change
When reconciliating manually, it may be possible to filter between
`All`, `Customers`, `Suppliers`, `Others`.

When changing this mode, we could end up in a situation such as:

1)

- reconciliation widget X is available
- reconciliation widget line XB is being loaded
- mode is changed -> reconciliation widget X is removed
- reconciliation widget line XB loading finishes

-> error because the parent reconciliation widget exists no more.

2)

having no (or less than the expected) or too few results because
removing reconciliation items (when for example we switch from supplier
to customer mode) can be done asynchronically. But the code expected it
to be done synchronically. Also all elements displayed were removed
instead of only the ones of other types.

thus instead of displaying the expected number of reconciliation items:
 MIN({num of record to display}, 10)

the records displayed were:
 MIN({num of record to display}, 10 - {num records displayed before})

which could amount to 0 event if there was thousands available ones.

This commit solves these two issues:

1) by not adding line to a reconciliation item without error,
2) by counting and removing displayed items correctly

opw-725796
opw-728309