[FIX] web: view_manager: reload views with correct params
Mode 'readonly' is a concept of list and form views. When
switching from a graph view to another view, then switching back
to graph, there was a crash because 'readonly' is not a known mode
of graph (in graph, mode is used to determine the type of graph to
display (pie, chart or bar)).

This is due to commit odoo/odoo@366053a which aims to reset the
mode to readonly when coming back to a form view using the
breadbrumbs.